+++ example/main.dart
@@ -6,11 +6,19 @@ void main() async {
 
   ///use Libretranslate
   final LibreTranslator = SimplyTranslator(EngineType.libre);
-/*
 
-  ///change instance (defaut is simplytranslate.org)
+  /// get the list with instances
+  print(GoogleTranslator.get_Instances);
+  //[simplytranslate.org, st.tokhmi.xyz, translate.josias.dev, ...
+
+  ///check if instance is working
+  print(await GoogleTranslator.is_instance_Working(
+      "simplytranslate.pussthecat.org"));
+  //true
+
   /// find other instances under https://simple-web.org/projects/simplytranslate.html
-  GoogleTranslator.baseUrl = "simplytranslate.pussthecat.org";
+  ///change instance (defaut is simplytranslate.org)
+  GoogleTranslator.set_Instance = "simplytranslate.pussthecat.org";
 
   ///if you do not specify the source language it is automatically selecting it depending on the text
   ///if you do not specify the target language it is automatically English
@@ -39,10 +47,11 @@ void main() async {
   print(Ltranslation.translations.text);
   //Die Anordnungen waren sehr kompliziert und schwierig.
 
-  ///without source language (auto):
+  ///without source language (auto) and choosing a random instance:
   Ltranslation = await LibreTranslator.translate(
       "The dispositions were very complicated and difficult.",
-      to: 'de');
+      to: 'de',
+      instanceMode: InstanceMode.Random);
   print(Ltranslation.translations.text);
   //Die Anordnungen waren sehr kompliziert und schwierig.
 
@@ -54,13 +63,13 @@ void main() async {
   textResult = await GoogleTranslator.tr("Er läuft schnell.");
   print(textResult);
   //He walks fast.
-*/
 
-  ///long form to also get definitions and single word translations
+  ///long form to also get definitions and single word translations and switching to next instance
   var Gtranslation = await GoogleTranslator.translate(
       "The dispositions were very complicated and difficult.",
       from: 'en',
-      to: 'de');
+      to: 'de',
+      instanceMode: InstanceMode.Loop);
 
   ///get whole Text translation
   ///Returns String
@@ -79,9 +88,26 @@ void main() async {
   print(Gtranslation.translations.raw_translations);
   //{adjective: {dick: {frequency: 1/3, words: [thick, fat, large, big, heavy, stout]}, faustdick: {frequency: 1/3,...
 
-  print("\n");
   ///get multiple word definitions in native language from Google
   ///returns Map<String, dynamic>
   print(Gtranslation.translations.raw_definitions);
   //{adjective: [{definition: of considerable size, extent, or intensity., synonyms: {: [large, sizeable,...
+
+  //get single word translations sorted by frequency
+  print(Gtranslation.translations.frequency_translations);
+  //[sehr, stark, bloß, genau, äußerste]
+
+  ///get Lists with Translations and Definitions
+  print(Gtranslation.translations.translations);
+  print(Gtranslation.translations.definitions);
+
+  ///changing instance each time
+  for (int i = 0; i < GoogleTranslator.get_Instances.length; i++) {
+    Gtranslation = await GoogleTranslator.translate("Gewechselt",
+        from: "de", instanceMode: InstanceMode.Random);
+    print(GoogleTranslator.get_current_Instance);
+    //translate.josias.dev
+    // translate.namazso.eu
+    // translate.riverside.rocks,...
+  }
 }