+++ lib/src/model/translation.dart
@@ -2,31 +2,42 @@ part of simplytranslate;
 
 ///Result includes the different translation results
 class Result {
-  const Result(this.text, this.raw_definitions, this.raw_translations, this.translations, this.definitions);
+  const Result(
+      this.text,
+      this.raw_definitions,
+      this.raw_translations,
+      this.translations,
+      this.definitions,
+      this.frequency_translations,
+      this.types);
   final String text;
   final Map<String, dynamic> raw_definitions;
   final Map<String, dynamic> raw_translations;
   final List<Translations> translations;
   final List<Definitions> definitions;
+  final List<String> frequency_translations;
+  final List<String> types;
 }
 
 ///Result includes the different translation results
 class Translations {
-  Translations(this.type,this.word, this.frequency, this.meaning);
+  Translations(this.type, this.word, this.frequency, this.meaning);
   final String type;
   final String word;
   final String frequency;
   final List<String> meaning;
 }
 
-class  Definitions{
-  Definitions(this.type,this.definition, this.synonyms, this.informal, this.use_in_sentence, this.archaic);
+class Definitions {
+  Definitions(this.type, this.definition, this.synonyms, this.informal_synonyms,
+      this.use_in_sentence, this.archaic, this.dictionary);
   final String type;
   final String definition;
   final List<String> synonyms;
-  final List<String> informal;
+  final List<String> informal_synonyms;
   final List<String> archaic;
   final String use_in_sentence;
+  final String dictionary;
 }
 
 /// Translation returned from SimplyTranslator.translate method